+++ eminus/xc/gga_x_pbe.py
@@ -90,10 +90,10 @@ def pbe_x_base(n, mu=0.2195149727645171, dn=None, **kwargs):
     kf = (3 * np.pi**2 * n) ** (1 / 3)
     # Handle divisions by zero
     # divkf = 1 / kf
-    divkf = np.divide(1, kf, out=np.zeros_like(kf), where=(kf > 0))
+    divkf = np.divide(1, kf, out=np.zeros_like(kf), where=kf > 0)
     # Handle divisions by zero
     # s = norm_dn * divkf / (2 * n)
-    s = np.divide(norm_dn * divkf, 2 * n, out=np.zeros_like(n), where=(n > 0))
+    s = np.divide(norm_dn * divkf, 2 * n, out=np.zeros_like(n), where=n > 0)
     f1 = 1 + mu * s**2 / kappa
     Fx = kappa - kappa / f1
     exunif = -3 * kf / (4 * np.pi)
@@ -109,6 +109,6 @@ def pbe_x_base(n, mu=0.2195149727645171, dn=None, **kwargs):
     # Handle divisions by zero
     # vsigmax = exunifdFx * divkf / (2 * norm_dn)
     vsigmax = np.divide(
-        exunifdFx * divkf, 2 * norm_dn, out=np.zeros_like(norm_dn), where=(norm_dn > 0)
+        exunifdFx * divkf, 2 * norm_dn, out=np.zeros_like(norm_dn), where=norm_dn > 0
     )
     return sx * n, np.array([vx]), vsigmax