+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ dev
 - Miscellaneous
    - Improve GDSMFB thermal functional
    - Drop Python 3.7 support
+   - Rework type hints
 
 v3.0.1 - Dec 18, 2024
 ---------------------