+++ CHANGELOG.rst
@@ -5,8 +5,8 @@
 Changelog
 =========
 
-dev
----
+v3.0.2 - Jan 16, 2025
+---------------------
 - New features
    - eminus paper release!
    - Add Jax powered FFT operators as an extra (thanks to @artemfilatov1)